requirements.txt
aiohttp                        3.8.1
aiosignal                      1.2.0
aiosqlite                      0.17.0
anyio                          3.6.1
appnope                        0.1.3
argon2-cffi                    21.3.0
argon2-cffi-bindings           21.2.0
asgiref                        3.5.2
asttokens                      2.0.5
async-generator                1.10
async-timeout                  4.0.2
atlastk                        0.13.2
attrs                          21.4.0
Babel                          2.10.3
backcall                       0.2.0
backports.zoneinfo             0.2.1
bcrypt                         3.2.2
beautifulsoup4                 4.11.1
bleach                         5.0.1
bot-studio                     1.3.9
certifi                        2022.5.18.1
cffi                           1.15.0
charset-normalizer             2.0.12
click                          8.1.3
cryptography                   37.0.2
cycler                         0.11.0
dateparser                     1.1.1
debugpy                        1.6.2
decorator                      5.1.1
defusedxml                     0.7.1
Deprecated                     1.2.13
elastic-transport              8.1.2
elasticsearch                  7.17.4
elasticsearch-async            6.2.0
elasticsearch-dsl              7.4.0
entrypoints                    0.4
et-xmlfile                     1.1.0
executing                      0.8.3
fastapi                        0.79.0
fastapi-csrf-protect           0.2.2
fastapi-jwt-auth               0.5.0
fastapi-sso                    0.5.1
fastjsonschema                 2.15.3
fonttools                      4.33.3
frozenlist                     1.3.0
gensim                         4.2.0
greenlet                       1.1.2
h11                            0.12.0
httmock                        1.4.0
httpcore                       0.15.0
httpx                          0.23.0
idna                           3.3
importlib-metadata             4.12.0
importlib-resources            5.8.0
influxdb                       5.3.1
influxdb-client                1.29.1
ipykernel                      6.15.1
ipython                        8.4.0
ipython-genutils               0.2.0
iso8601                        1.0.2
itsdangerous                   2.1.2
jedi                           0.18.1
Jinja2                         3.1.2
json5                          0.9.8
jsonschema                     4.7.2
jupyter-client                 7.3.4
jupyter-core                   4.11.1
jupyter-server                 1.18.1
jupyterlab                     3.4.3
jupyterlab-pygments            0.2.2
jupyterlab-server              2.15.0
kiwisolver                     1.4.3
limits                         2.6.3
MarkupSafe                     2.1.1
matplotlib                     3.5.2
matplotlib-inline              0.1.3
mistune                        0.8.4
msgpack                        1.0.4
multidict                      6.0.2
mysql-connector-python         8.0.29
mysqlclient                    2.1.1
nbclassic                      0.4.3
nbclient                       0.6.6
nbconvert                      6.5.0
nbformat                       5.4.0
nest-asyncio                   1.5.5
notebook                       6.4.12
notebook-shim                  0.1.0
numpy                          1.23.1
oauthlib                       3.2.0
openpyxl                       3.0.10
outcome                        1.2.0
packaging                      21.3
pandas                         1.4.3
pandocfilters                  1.5.0
parso                          0.8.3
passlib                        1.7.4
pexpect                        4.8.0
pickleshare                    0.7.5
Pillow                         9.1.1
pip                            22.1.2
prometheus-api-client          0.5.1
prometheus-client              0.14.1
prompt-toolkit                 3.0.30
protobuf                       4.21.2
psutil                         5.9.1
ptyprocess                     0.7.0
pure-eval                      0.2.2
pyasn1                         0.4.8
pycparser                      2.21
pydantic                       1.9.1
Pygments                       2.12.0
PyJWT                          1.7.1
pylogbeat                      2.0.0
pyOpenSSL                      22.0.0
pyparsing                      3.0.9
pypika-tortoise                0.1.5
pyrsistent                     0.18.1
PySocks                        1.7.1
python-dateutil                2.8.2
python-dotenv                  0.20.0
python-logstash-async          2.5.0
python-multipart               0.0.5
pytz                           2022.1
pytz-deprecation-shim          0.1.0.post0
pyzmq                          23.2.0
regex                          2022.3.2
requests                       2.27.1
rfc3986                        1.5.0
rsa                            4.8
Rx                             3.2.0
scipy                          1.8.1
selenium                       4.2.0
Send2Trash                     1.8.0
setuptools                     49.2.1
six                            1.16.0
smart-open                     6.0.0
sniffio                        1.2.0
sortedcontainers               2.4.0
soupsieve                      2.3.2.post1
SQLAlchemy                     1.4.38
stack-data                     0.3.0
starlette                      0.19.1
terminado                      0.15.0
tinycss2                       1.1.1
tornado                        6.2
tortoise                       0.1.1
tortoise-orm                   0.19.1
tqdm                           4.64.0
traitlets                      5.3.0
trio                           0.21.0
trio-websocket                 0.9.2
typing_extensions              4.2.0
tzdata                         2022.1
tzlocal                        4.2
urllib3                        1.26.9
uvicorn                        0.18.2
uvloop                         0.16.0
wcwidth                        0.2.5
webdriver-manager              3.7.0
webencodings                   0.5.1
websocket-client               1.3.3
wrapt                          1.14.1
wsproto                        1.1.0
yarl                           1.7.2
youtube-comment-scraper-python 1.0.0
zipp                           3.8.1